requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
xgboost
lightgbm
joblib
hyperopt
optuna
pydantic
parquet
python-dotenv
fastparquet